bonus + amelioration

+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: cllovio <cllovio@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 13:53:54 by cllovio           #+#    #+#             */
-/*   Updated: 2022/11/12 11:13:44 by cllovio          ###   ########.fr       */
+/*   Updated: 2022/11/21 13:44:53 by cllovio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, unsigned int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*bb;
 
 	i = 0;
